requirements.txt
beautifulsoup4==4.12.3
python-telegram-bot==21.0.1
requests==2.31.0
python-dotenv==1.0.1
lxml==5.2.1
